Actualizacion de la base de datos, borrado multiple y ajustes del grafico

+++ src/Controller/TareasController.php
@@ -13,6 +13,8 @@ use Symfony\Component\Security\Http\Attribute\IsGranted;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Bundle\SecurityBundle\Security;
 
 #[Route('/tareas')]
 #[IsGranted('ROLE_USER')]
@@ -34,12 +36,12 @@ class TareasController extends AbstractController
     public function index(TareasRepository $tareasRepository): Response
     {
         return $this->render('tareas/index.html.twig', [
-            'tareas' => $tareasRepository->findAll(),
+            'tareas' => $tareasRepository->findNotDeleted(),
         ]);
     }
 
     #[Route('/new', name: 'app_tareas_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    public function new(Request $request, EntityManagerInterface $entityManager, Security $security): Response
     {
         $tarea = new Tareas();
         $proyecto = new Proyectos();
@@ -49,6 +51,11 @@ class TareasController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $tarea->setCreadoPor($security->getUser());
+            $tarea->setActualizadoPor($security->getUser());
+            $tarea->setActualizadoEn(new \DateTime("now", new \DateTimeZone('Europe/Madrid')));
+            $tarea->setCreadoEn(new \DateTime("now", new \DateTimeZone('Europe/Madrid')));
+
             $entityManager->persist($tarea);
             $entityManager->flush();
 
@@ -65,17 +72,16 @@ class TareasController extends AbstractController
             // Persistir el nuevo proyecto
             $entityManager->persist($proyecto);
             $entityManager->flush();
-    
+
             // Obtener la URL de la página de crear nueva tarea
             $urlNuevaTarea = $this->generateUrl('app_tareas_new');
-    
+
             // Devolver la URL de la página de crear nueva tarea en formato JSON
             return new JsonResponse(['urlNuevaTarea' => $urlNuevaTarea]);
         }
-    
+
         // En caso de error, devolver una respuesta de error
         return new JsonResponse(['error' => 'Error al guardar el nuevo proyecto'], Response::HTTP_BAD_REQUEST);
-    
     }
 
     #[Route('/{id}', name: 'app_tareas_show', methods: ['GET'])]
@@ -93,56 +99,36 @@ class TareasController extends AbstractController
         $form->handleRequest($request);
         $proyectoForm = $this->createForm(ProyectosType::class);
         $proyectoForm->handleRequest($request);
-         // Obtener el proyecto asociado a la tarea
+        // Obtener el proyecto asociado a la tarea
         $proyecto = $tarea->getProyecto();
-    
+
         if ($form->isSubmitted() && $form->isValid()) {
             // Asignar el proyecto seleccionado a la tarea
             $proyectoSeleccionado = $form->get('proyecto')->getData();
             $tarea->setProyecto($proyectoSeleccionado);
-    
+
             // Guardar la tarea
             $entityManager->flush();
-    
+
             return $this->redirectToRoute('app_tareas_index', [], Response::HTTP_SEE_OTHER);
         }
-    
+
         return $this->render('tareas/edit.html.twig', [
             'tarea' => $tarea,
             'form' => $form->createView(),
             'proyectoForm' => $proyectoForm->createView(),
         ]);
     }
-    
-    
 
-    #[Route('/{id}', name: 'app_tareas_delete', methods: ['POST'])]
-    public function delete(Request $request, Tareas $tarea, EntityManagerInterface $entityManager): Response
-    {
-        $id = $request->attributes->get('id');
-        $tarea = $this->tareasRepository->find($id);
-
-        if (!$tarea) {
-            throw $this->createNotFoundException('Tarea no encontrado');
-        }
-
-        if ($this->isCsrfTokenValid('delete'.$tarea->getId(), $request->request->get('_token'))) {
-            $entityManager->remove($tarea);
-            $entityManager->flush();
-        }
-
-        return $this->redirectToRoute('app_tareas_index');
-    }
 
     #[Route('/eliminar_tareas', name: 'eliminar_tareas', methods: ['POST'])]
-    public function eliminarTareas(Request $request, EntityManagerInterface $entityManager)
+    public function eliminarTareas(Request $request, EntityManagerInterface $entityManager, Security $security)
     {
-
         // Decodificar el contenido JSON de la solicitud
         $data = json_decode($request->getContent(), true);
         $tareasIds = $data['ids'];
 
-        // Obtener el EntityManager
+        $usuarioActual = $security->getUser(); // Obtener el usuario actual con Symfony Security
 
         foreach ($tareasIds as $tareaId) {
             // Buscar la tarea por su ID
@@ -152,15 +138,36 @@ class TareasController extends AbstractController
                 return new JsonResponse(['error' => 'Tarea no encontrada con el ID: ' . $tareaId], 404);
             }
 
-            // Eliminar la tarea
-            $entityManager->remove($tarea);
+            // Realizar borrado lógico
+            $tarea->setEliminado(true);
+            $tarea->setActualizadoPor($usuarioActual); // Asignar el usuario que hace la actualización
+            $tarea->setActualizadoEn(new \DateTime("now", new \DateTimeZone('Europe/Madrid'))); // Asignar la fecha actual
+
+            $entityManager->persist($tarea); // Esto puede no ser necesario dependiendo de la configuración de Doctrine
         }
 
         // Guardar los cambios en la base de datos
         $entityManager->flush();
 
-        return new JsonResponse(['message' => 'Tareas eliminadas correctamente.']);
+        return new JsonResponse(['message' => 'Tareas actualizadas como eliminadas correctamente.']);
+    }
+
+
+    #[Route('/delete/{id}', name: 'app_tareas_delete', methods: ['POST'])]
+    public function delete(Request $request, Tareas $tarea, EntityManagerInterface $entityManager): Response
+    {
+        $id = $request->attributes->get('id');
+        $tarea = $this->tareasRepository->find($id);
+
+        if (!$tarea) {
+            throw $this->createNotFoundException('Tarea no encontrado');
+        }
+
+        if ($this->isCsrfTokenValid('delete' . $tarea->getId(), $request->request->get('_token'))) {
+            $entityManager->remove($tarea);
+            $entityManager->flush();
+        }
+
+        return $this->redirectToRoute('app_tareas_index');
     }
 }
-
-
